pyproject.toml
[project]
name = "gaze-vit-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "einops>=0.8.1",
    "gymnasium[atari]>=1.2.3",
    "hydra-core>=1.3.2",
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.8",
    "moviepy>=2.2.1",
    "numpy>=2.4.1",
    "opencv-python>=4.11.0.86",
    "pygame>=2.6.1",
    "scipy>=1.17.0",
    "stable-baselines3>=2.3.2",
    "tensorboard>=2.20.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "vit-pytorch",
]

[tool.uv.sources]
vit-pytorch = { git = "ssh://git@github.com/lucidrains/vit-pytorch.git" }
